refactor: routes.ts file config without file routing
